+++ src/main/java/seedu/duke/ChatFxml/DialogBox.java
@@ -1,0 +1,69 @@
+package seedu.duke.chatfxml;
+
+import javafx.fxml.FXML;
+import javafx.fxml.FXMLLoader;
+import java.util.Collections;
+import javafx.geometry.Pos;
+import javafx.scene.control.Label;
+import javafx.scene.image.ImageView;
+import javafx.scene.image.Image;
+import javafx.scene.layout.HBox;
+import javafx.collections.FXCollections;
+import javafx.collections.ObservableList;
+import javafx.scene.Node;
+import java.io.IOException;
+
+public class DialogBox extends HBox {
+
+    @FXML
+    private Label dialog;
+    @FXML
+    private ImageView displayPicture;
+
+    private DialogBox(String text, Image img) {
+        try {
+            FXMLLoader fxmlLoader = new FXMLLoader(MainWindow.class.getResource("/view/DialogBox.fxml"));
+            fxmlLoader.setController(this);
+            fxmlLoader.setRoot(this);
+            fxmlLoader.load();
+        } catch (IOException e) {
+            e.printStackTrace();
+        }
+
+        dialog.setText(text);
+        displayPicture.setImage(img);
+        this.setAlignment(Pos.TOP_RIGHT);
+    }
+
+    /**
+     * Flips the dialog box such that the Image is on the left and text on the right.
+     */
+    private void flip() {
+        this.setAlignment(Pos.TOP_LEFT);
+        ObservableList<Node> tmp = FXCollections.observableArrayList(this.getChildren());
+        Collections.reverse(tmp);
+        this.getChildren().setAll(tmp);
+    }
+
+    /**
+     * Creates a dialog box containing the users profile picture and the user's message.
+     * @param text text user has inputted.
+     * @param img user profile picture.
+     * @return dialog box containing the user's message and the users profile pic on the right.
+     */
+    public static DialogBox getUserDialog(String text, Image img) {
+        return new DialogBox(text, img);
+    }
+
+    /**
+     * Creates a dialog box containing the duke system's profile picture and its response user's message.
+     * @param text text user has inputted.
+     * @param img duke system's profile picture.
+     * @return dialog box containing the system's response to the user's message and with its profile pic on the left.
+     */
+    public static DialogBox getDukeDialog(String text, Image img) {
+        var db = new DialogBox(text, img);
+        db.flip();
+        return db;
+    }
+}
