+++ src/main/java/seedu/duke/ChatFxml/MainWindow.java
@@ -1,21 +1,13 @@
-package seedu.duke.ChatFxml;
+package seedu.duke.chatfxml;
 
 import javafx.fxml.FXML;
-import javafx.fxml.FXMLLoader;
 import javafx.scene.control.Button;
-import javafx.scene.Scene;
 import javafx.scene.control.ScrollPane;
 import javafx.scene.control.TextField;
 import javafx.scene.layout.AnchorPane;
 import javafx.scene.layout.VBox;
-import javafx.stage.Stage;
-import javafx.scene.control.Label;
 import javafx.scene.image.Image;
-import javafx.scene.image.ImageView;
-import java.io.IOException;
-import seedu.duke.DialogBox;
 import seedu.duke.Duke;
-import seedu.duke.Main;
 
 /**
  * Controller for MainWindow. Provides the layout for the other controls.
@@ -34,24 +26,15 @@ public class MainWindow extends AnchorPane {
     private Image dukePic = new Image(this.getClass().getClassLoader().getResourceAsStream("images/DaDuke.png"));
     private Duke duke = new Duke();
 
+    /**
+     * Initialize the main window for the chat between user and duke system.
+     */
     @FXML
     public void initialize() {
         scrollPane.vvalueProperty().bind(dialogContainer.heightProperty());
         dialogContainer.getChildren().addAll(
-                DialogBox.getDukeDialog(new Label(duke.getWelcomeMessage()), new ImageView(dukePic))
+                DialogBox.getDukeDialog(duke.getWelcomeMessage(), dukePic)
         );
-    }
-
-    public void start(Stage stage) {
-        try {
-            FXMLLoader fxmlLoader = new FXMLLoader(Main.class.getClassLoader().getResource("view/MainWindow.fxml"));
-            AnchorPane ap = (AnchorPane) fxmlLoader.load();
-            Scene scene = new Scene(ap);
-            stage.setScene(scene);
-            stage.show();
-        } catch (IOException e) {
-            e.printStackTrace();
-        }
     }
 
     /** Creates two dialog boxes, one echoing user input and the other containing Duke's reply and then appends them to
@@ -62,8 +45,8 @@ public class MainWindow extends AnchorPane {
         String input = userInput.getText();
         String response = duke.getResponse(input);
         dialogContainer.getChildren().addAll(
-                //DialogBox.getUserDialog(input, user),
-                //DialogBox.getDukeDialog(response, duke)
+                DialogBox.getUserDialog(input, userPic),
+                DialogBox.getDukeDialog(response, dukePic)
         );
         userInput.clear();
     }
