--- src/main/java/seedu/duke/DialogBox.java
@@ -1,56 +1,0 @@
-package seedu.duke;
-
-import java.util.Collections;
-import javafx.geometry.Pos;
-import javafx.scene.control.Label;
-import javafx.scene.image.ImageView;
-import javafx.scene.layout.Background;
-import javafx.scene.layout.BackgroundFill;
-import javafx.scene.layout.HBox;
-import javafx.collections.FXCollections;
-import javafx.collections.ObservableList;
-import javafx.scene.Node;
-import javafx.geometry.Insets;
-import javafx.scene.paint.Color;
-import javafx.scene.layout.CornerRadii;
-
-public class DialogBox extends HBox {
-
-    private Label text;
-    private ImageView displayPicture;
-    public DialogBox(Label l, ImageView iv) {
-
-        text = l;
-        text.setPadding(new Insets(10,20,10,20));
-        text.setWrapText(true);
-
-        displayPicture = iv;
-        displayPicture.setFitWidth(100.0);
-        displayPicture.setFitHeight(100.0);
-
-        this.setAlignment(Pos.TOP_RIGHT);
-        this.getChildren().addAll(text, displayPicture);
-        this.setPadding(new Insets(10,20,10,10));
-        this.setBackground(new Background(new BackgroundFill(Color.rgb(246, 243, 99), CornerRadii.EMPTY, Insets.EMPTY)));
-    }
-
-    /**
-     * Flips the dialog box such that the ImageView is on the left and text on the right.
-     */
-    private void flip() {
-        this.setAlignment(Pos.TOP_LEFT);
-        ObservableList<Node> tmp = FXCollections.observableArrayList(this.getChildren());
-        Collections.reverse(tmp);
-        this.getChildren().setAll(tmp);
-    }
-
-    public static DialogBox getUserDialog(Label l, ImageView iv) {
-        return new DialogBox(l, iv);
-    }
-
-    public static DialogBox getDukeDialog(Label l, ImageView iv) {
-        var db = new DialogBox(l, iv);
-        db.flip();
-        return db;
-    }
-}
